# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ColaboraCar API
  version: 0.1.0
  summary: Contrato entre el frontend (React) y el backend (Spring Boot) de ColaboraCar.
  description: |
    Contrato de la API REST de ColaboraCar para el alcance actual (RF-01 a RF-22 y
    RN-01 a RN-16 del `README.md`): autenticación, registro, catálogos, archivos,
    vehículo, viajes, reservas y perfil.

    ## Convenciones

    - **Campos JSON**: inglés en `camelCase`. El backend traduce hacia y desde los
      documentos de MongoDB, que están en español y `snake_case` (cada schema indica su
      colección de origen).
    - **Identificadores**: cadenas de 24 caracteres hexadecimales (ObjectId de MongoDB).
    - **Fechas**: ISO 8601 en UTC (`2026-10-09T13:30:00Z`). El frontend las muestra en la
      hora de Perú (`America/Lima`, UTC-5).
    - **Coordenadas**: objeto `{ lat, lng }` en grados decimales. El backend las convierte
      al `Point` GeoJSON (`[lng, lat]`) que se guarda en MongoDB.
    - **Créditos**: enteros. En este alcance solo se muestran y se registran; no hay cobro
      ni saldo.
    - **Listas**: se devuelven dentro de `{ "data": [...] }`. Un recurso individual se
      devuelve directamente, sin envoltorio.
    - **Errores**: toda respuesta 4xx o 5xx tiene el cuerpo `Error`:
      `{ code, message, errors?, details? }`. `message` va en español y se puede mostrar
      al usuario; `code` es estable y es lo que el frontend debe usar para decidir.
...

    ## Autenticación

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /vehicles:
    post:
      tags: [Vehicles]
      operationId: registerVehicle
      summary: Registrar el vehículo
      description: |
        Guarda el vehículo dentro del usuario y le agrega el rol `DRIVER`. Admite un token
        `REGISTRATION` (durante el registro) o `SESSION` (un pasajero que registra su
        vehículo más adelante). El perfil debe estar completo. Un usuario solo puede tener
        un vehículo.

        Todos los campos son obligatorios, excepto la casilla «Soy el propietario» y la
        foto (RN-06).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreateRequest"
      responses:
        "201":
          description: Vehículo registrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          description: |
            Datos del vehículo inválidos (`VALIDATION_ERROR`; `errors` detalla cada campo)
            o foto inexistente, de otro usuario o de otro propósito
            (`INVALID_FILE_REFERENCE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VALIDATION_ERROR
                message: Revisa los datos enviados
                errors:
                  - field: plate
                    message: La placa debe tener el formato ABC-123
                  - field: ownerDni
                    message: El DNI debe tener 8 dígitos
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: |
            La placa pertenece a otro usuario (`PLATE_ALREADY_REGISTERED`), el usuario ya
            tiene un vehículo (`VEHICLE_ALREADY_REGISTERED`) o aún no completó su perfil
            (`PROFILE_INCOMPLETE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: PLATE_ALREADY_REGISTERED
                message: La placa ya está registrada
  /vehicles/me:
    get:
      tags: [Vehicles]
      operationId: getMyVehicle
      summary: Consultar mi vehículo
      description: |
        Lo usa «Publicar un viaje» para saber si el usuario puede publicar (RF-18).
      responses:
        "200":
          description: Vehículo del usuario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: El usuario no tiene vehículo registrado (`VEHICLE_NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VEHICLE_NOT_FOUND
                message: No tienes un vehículo registrado
    put:
      tags: [Vehicles]
      operationId: updateMyVehicle
      summary: Actualizar mi vehículo
      description: |
        Reemplaza los datos editables del vehículo y actualiza la copia en los viajes del
        conductor en estado `PUBLISHED` o `IN_PROGRESS`. `ownerDni` no se modifica aquí.

        Si el usuario todavía no tiene vehículo, lo registra en lugar de rechazar la
        solicitud: guarda el vehículo, le agrega el rol `DRIVER` y no hay viajes que
        actualizar. En ese caso el perfil debe estar completo y `ownerDni` es opcional: si
        no llega, se usa el DNI del usuario, si lo registró.

        Si `seats` baja, las plazas de los viajes ya publicados no cambian.

        Todos los datos son obligatorios, incluida la foto (RN-15): si el vehículo aún no
        tiene foto, `photoFileId` es obligatorio. De las casillas, solo es obligatoria la
        de términos y condiciones (RN-16).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdateRequest"
      responses:
        "200":
          description: Vehículo actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          description: |
            Datos del vehículo inválidos (`VALIDATION_ERROR`; `errors` detalla cada campo)
            o foto inexistente, de otro usuario o de otro propósito
            (`INVALID_FILE_REFERENCE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VALIDATION_ERROR
                message: Revisa los datos enviados
                errors:
                  - field: seats
                    message: Las plazas deben estar entre 1 y 10
                  - field: brand
                    message: La marca es obligatoria
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: |
            La placa pertenece a otro usuario (`PLATE_ALREADY_REGISTERED`) o el usuario no
            tiene vehículo y aún no completó su perfil (`PROFILE_INCOMPLETE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rides:
    get:
      tags: [Rides]
      operationId: listRides
      summary: Listar los viajes disponibles
      description: |
        Devuelve los viajes en estado `PUBLISHED` de una sede, con salida futura y con
        plazas disponibles, ordenados por `departureTime` ascendente. No incluye los
        viajes publicados por el propio usuario.

        La búsqueda siempre es por sede: solo se listan los viajes publicados por
        alumnos de esa sede. Sin `campusId` se usa la sede del usuario autenticado.

        Cada elemento trae `conditions`, de modo que el modal «Condiciones del viaje» no
        necesita otra petición.
      parameters:
        - name: campusId
          in: query
          required: false
          description: |
            Sede cuyos viajes se listan. Si no se envía, se usa la sede del usuario
            autenticado. Si la sede no existe, o el usuario aún no tiene sede, la lista
            llega vacía.
          schema:
            $ref: "#/components/schemas/ObjectId"
        - name: destination
          in: query
          required: false
          description: |
            Texto a buscar en la etiqueta o la dirección del destino, sin distinguir
            mayúsculas ni tildes.
          schema:
            type: string
            maxLength: 100
          example: Huaca del Dragón
        - name: time
          in: query
          required: false
          description: |
            Hora de salida mínima, en formato `HH:mm` de 24 horas y hora de Perú. Devuelve
            los viajes de hoy que salen a esa hora o después.
          schema:
            type: string
            pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "14:15"
        - name: passengers
          in: query
          required: false
          description: Plazas que se necesitan; filtra por `availableSeats >= passengers`.
          schema:
            type: integer
            minimum: 1
            maximum: 99
            default: 1
      responses:
        "200":
          description: Viajes disponibles.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RideSummary"
        "400":
          description: Algún filtro tiene un formato inválido (`VALIDATION_ERROR`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VALIDATION_ERROR
                message: Revisa los datos enviados
                errors:
                  - field: time
                    message: La hora debe tener el formato HH:mm
                  - field: passengers
                    message: El número de pasajeros debe ser al menos 1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Rides]
      operationId: publishRide
      summary: Publicar un viaje
      description: |
        Crea el viaje en estado `PUBLISHED`; queda visible de inmediato en el listado.
        Solo para usuarios con vehículo registrado (RF-18).

        El backend toma del usuario autenticado el conductor, el vehículo y la sede, y
        calcula `distanceKm` (línea recta entre origen, paradas y destino) y
        `durationMin` (estimación a partir de la distancia). El frontend no los envía.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideCreateRequest"
      responses:
        "201":
          description: Viaje publicado.
          headers:
            Location:
              description: Ruta del viaje creado (`/api/rides/{rideId}`).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideDetail"
        "400":
          description: |
            Datos inválidos (`VALIDATION_ERROR`): por ejemplo, `departureTime` en el
            pasado o fuera del horario de 6:00 a 23:00, `pricePerSeat` fuera del rango de
            1 a 10, `seats` mayor que las plazas del vehículo o más de 5 paradas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VALIDATION_ERROR
                message: Revisa los datos enviados
                errors:
                  - field: departureTime
                    message: La hora de salida debe ser futura y estar entre las 6:00 y las 23:00
                  - field: pricePerSeat
                    message: El precio por plaza debe estar entre 1 y 10
                  - field: seats
                    message: Las plazas no pueden superar las de tu vehículo
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            El usuario no tiene vehículo registrado (`VEHICLE_REQUIRED`) o el token no
            es de sesión (`FORBIDDEN_SCOPE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VEHICLE_REQUIRED
                message: Registra un vehículo para publicar viajes
  /rides/{rideId}:
    parameters:
      - $ref: "#/components/parameters/RideId"
    get:
      tags: [Rides]
      operationId: getRide
      summary: Consultar el detalle de un viaje
      description: |
        Devuelve el viaje con la ruta completa (origen, paradas y destino con
        coordenadas) para dibujar el mapa. Responde para cualquier estado del viaje.
      responses:
        "200":
          description: Detalle del viaje.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RideNotFound"
  /rides/{rideId}/reserve:
    parameters:
      - $ref: "#/components/parameters/RideId"
    post:
      tags: [Rides]
      operationId: reserveRide
      summary: Reservar un viaje
      description: |
        Descuenta las plazas de forma atómica y crea la reserva en estado `CONFIRMED`,
        todo en una misma transacción. Un pasajero solo puede tener una reserva
        `CONFIRMED` por viaje.

        El cuerpo es opcional: sin cuerpo se reserva 1 plaza.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveRideRequest"
      responses:
        "201":
          description: Reserva confirmada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: Reserva confirmada
        "400":
          description: El valor de `seats` no es un entero mayor o igual a 1 (`VALIDATION_ERROR`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: VALIDATION_ERROR
                message: Revisa los datos enviados
                errors:
                  - field: seats
                    message: Debes reservar al menos 1 plaza
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            El conductor no puede reservar su propio viaje (`OWN_RIDE`) o el token no es
            de sesión (`FORBIDDEN_SCOPE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/RideNotFound"
        "409":
          description: |
            No quedan plazas suficientes (`NO_SEATS_AVAILABLE`), el usuario ya tiene una
            reserva confirmada en este viaje (`ALREADY_BOOKED`) o el viaje ya no está
            publicado o ya salió (`RIDE_NOT_AVAILABLE`). En los tres casos el frontend
            debe refrescar el listado. También si el usuario aún no completó su perfil
            (`PROFILE_INCOMPLETE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: NO_SEATS_AVAILABLE
                message: Este viaje ya no tiene asientos disponibles
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT con el claim `scope` (`REGISTRATION`, `PRE_AUTH` o `SESSION`). Salvo que el
        endpoint indique otra cosa, se exige `SESSION`.
  parameters:
    RideId:
      name: rideId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ObjectId"
  responses:
    Unauthorized:
      description: Falta el token, es inválido o expiró (`UNAUTHORIZED`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: UNAUTHORIZED
            message: Tu sesión expiró. Vuelve a iniciar sesión.
    Forbidden:
      description: El token es válido, pero su `scope` no admite este endpoint (`FORBIDDEN_SCOPE`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: FORBIDDEN_SCOPE
            message: No tienes permiso para realizar esta acción
    RideNotFound:
      description: El viaje no existe (`RIDE_NOT_FOUND`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: RIDE_NOT_FOUND
            message: Viaje no encontrado
  schemas:
...
    ObjectId:
      type: string
      pattern: "^[0-9a-f]{24}$"
      description: Identificador de MongoDB (ObjectId) como cadena.
      example: 6705a1f0c3d4e5f600000501
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Código estable del error. Valores: `VALIDATION_ERROR`, `UNAUTHORIZED`,
            `FORBIDDEN_SCOPE`, `INVALID_CREDENTIALS`, `ACCOUNT_BLOCKED`,
            `EMAIL_ALREADY_REGISTERED`, `INVALID_CATALOG_REFERENCE`, `PROFILE_INCOMPLETE`,
            `INVALID_FILE_REFERENCE`, `FILE_TOO_LARGE`, `UNSUPPORTED_FILE_TYPE`,
            `OTP_INVALID`, `OTP_EXPIRED`, `OTP_ATTEMPTS_EXCEEDED`, `OTP_RESEND_TOO_SOON`,
            `PLATE_ALREADY_REGISTERED`, `VEHICLE_ALREADY_REGISTERED`, `VEHICLE_NOT_FOUND`,
            `VEHICLE_REQUIRED`, `RIDE_NOT_FOUND`, `RIDE_NOT_AVAILABLE`,
            `NO_SEATS_AVAILABLE`, `ALREADY_BOOKED`, `OWN_RIDE`, `INTERNAL_ERROR`.
          example: VALIDATION_ERROR
        message:
          type: string
          description: |
            Mensaje en español, apto para mostrarse al usuario. Con `VALIDATION_ERROR` y
            `errors` es siempre el mensaje general `Revisa los datos enviados`: el detalle
            de cada campo va en `errors`.
        errors:
          type: array
          description: |
            Errores por campo, uno por cada campo o parámetro inválido. Solo con
            `VALIDATION_ERROR`.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                description: Nombre del campo; con notación de punto si está anidado (`origin.location.lat`).
              message:
                type: string
        details:
          type: object
          description: |
            Datos adicionales de algunos errores: `attemptsLeft` en `OTP_INVALID` y
            `retryAfter` (segundos) en `OTP_RESEND_TOO_SOON`.
          additionalProperties: true
...
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: -8.0975
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -79.0353
    Place:
      type: object
      description: Punto de una ruta o dirección (`etiqueta`, `direccion`, `ubicacion`).
      required: [label, location]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 80
          example: Universidad Tecnológica del Perú
        address:
          type: [string, "null"]
          maxLength: 160
          example: Sede Trujillo, Av. Nicolás de Piérola
        location:
          $ref: "#/components/schemas/GeoPoint"
...
    PhotoFileId:
      allOf:
        - $ref: "#/components/schemas/ObjectId"
      description: |
        `id` devuelto por `POST /files`. La imagen debe ser del usuario autenticado y del
        propósito que corresponde al formulario.
...
    VehicleType:
      type: string
      enum: [SEDAN, HATCHBACK, SUV, VAN, MOTORCYCLE]
      description: En MongoDB, `MOTORCYCLE` es `MOTO`; el resto se llama igual.
    VehicleStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
      description: En MongoDB, `ACTIVO` e `INACTIVO`.
    Vehicle:
      type: object
      description: Campo embebido `usuarios.vehiculo`.
      required:
        - plate
        - type
        - brand
        - model
        - color
        - year
        - seats
        - ownerDni
        - isOwner
        - status
        - photoUrl
        - registeredAt
      properties:
        plate:
          type: string
          example: ABC-123
        type:
          $ref: "#/components/schemas/VehicleType"
        brand:
          type: string
          example: Toyota
        model:
          type: string
          example: Yaris
        color:
          type: string
          example: Blanco
        year:
          type: integer
          example: 2021
        seats:
          type: integer
          description: Plazas que el conductor comparte, sin contar la suya.
          example: 3
        ownerDni:
          type: [string, "null"]
          description: "`null` si el vehículo se registró desde «Actualizar vehículo» sin DNI del propietario."
          example: "45678912"
        isOwner:
          type: boolean
        status:
          $ref: "#/components/schemas/VehicleStatus"
        photoUrl:
          type: [string, "null"]
          description: "`null` si el vehículo se registró sin foto (es opcional en el registro, RN-06)."
        registeredAt:
          type: string
          format: date-time
    VehicleCreateRequest:
      type: object
      required:
        [plate, type, brand, model, color, year, seats, ownerDni, acceptedTerms]
      properties:
        plate:
          $ref: "#/components/schemas/Plate"
        type:
          $ref: "#/components/schemas/VehicleType"
        year:
          $ref: "#/components/schemas/VehicleYear"
        seats:
          $ref: "#/components/schemas/VehicleSeats"
        ownerDni:
          type: string
          pattern: '^\d{8}$'
          example: "45678912"
        isOwner:
          type: boolean
          default: false
          description: Casilla «Soy el propietario del vehículo». Es opcional.
        acceptedTerms:
          type: boolean
          const: true
          description: Debe ser `true`; el backend guarda la fecha de aceptación.
        brand:
          type: string
          minLength: 1
          maxLength: 40
          example: Toyota
        model:
          type: string
          minLength: 1
          maxLength: 40
          example: Yaris
        color:
          type: string
          minLength: 1
          maxLength: 30
          example: Blanco
        photoFileId:
          allOf:
            - $ref: "#/components/schemas/PhotoFileId"
          description: Foto del vehículo. Es opcional en el registro.
    VehicleUpdateRequest:
      type: object
      required: [plate, type, brand, model, color, year, seats, acceptedTerms]
      properties:
        plate:
          $ref: "#/components/schemas/Plate"
        type:
          $ref: "#/components/schemas/VehicleType"
        brand:
          type: string
          minLength: 1
          maxLength: 40
          example: Toyota
        model:
          type: string
          minLength: 1
          maxLength: 40
          example: Yaris
        color:
          type: string
          minLength: 1
          maxLength: 30
          example: Blanco
        photoFileId:
          allOf:
            - $ref: "#/components/schemas/PhotoFileId"
          description: |
            Foto nueva. Si no se envía, se conserva la actual; es obligatorio si el
            vehículo aún no tiene foto.
        year:
          $ref: "#/components/schemas/VehicleYear"
        seats:
          $ref: "#/components/schemas/VehicleSeats"
        ownerDni:
          type: string
          pattern: '^\d{8}$'
          description: |
            Opcional. Solo se usa si el usuario aún no tiene vehículo y la actualización lo
            registra; si el vehículo ya existe, se ignora.
          example: "45678912"
        isOwner:
          type: boolean
          description: |
            Casilla «Soy el propietario legal o conductor autorizado». Es opcional; si no
            se envía, se conserva el valor actual (`false` si el vehículo es nuevo).
        acceptedTerms:
          type: boolean
          const: true
    Plate:
      type: string
      pattern: "^[A-Za-z0-9]{3}-?[A-Za-z0-9]{3,4}$"
      description: |
        Placa con el formato del Perú (RN-07). El backend la guarda en mayúsculas y con guion (`ABC-123`); es
        única entre todos los usuarios.
      example: ABC-123
    VehicleYear:
      type: integer
      minimum: 2000
      description: Año de fabricación; como mínimo 2000 y como máximo el año actual (RN-08).
      example: 2021
    VehicleSeats:
      type: integer
      minimum: 1
      maximum: 10
      description: Plazas que el conductor comparte, sin contar la suya; entre 1 y 10 (RN-09).
      example: 3

    # ----------------------------------------------------------------- Viajes
    RideDirection:
      type: string
      enum: [TO_CAMPUS, TO_HOME]
      description: |
        Sentido del viaje. En MongoDB, `IDA_UNIVERSIDAD` y `REGRESO_CASA`. En
        `TO_CAMPUS` el destino es la sede; en `TO_HOME`, el origen.
    RideStatus:
      type: string
      enum: [PUBLISHED, IN_PROGRESS, COMPLETED, CANCELLED]
      description: En MongoDB, `PUBLICADO`, `EN_CURSO`, `COMPLETADO` y `CANCELADO`.
...
    RideBase:
      type: object
      required:
        - id
        - status
        - direction
        - departureTime
        - pricePerSeat
        - totalSeats
        - availableSeats
        - distanceKm
        - durationMin
        - driver
        - vehicle
        - conditions
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        status:
          $ref: "#/components/schemas/RideStatus"
        direction:
          $ref: "#/components/schemas/RideDirection"
        departureTime:
          type: string
          format: date-time
          example: "2026-10-09T13:30:00Z"
        pricePerSeat:
          type: integer
          minimum: 1
          maximum: 10
          description: Aporte por plaza, en créditos.
          example: 5
        totalSeats:
          type: integer
          example: 3
        availableSeats:
          type: integer
          example: 3
        distanceKm:
          type: number
          description: Distancia en línea recta entre origen, paradas y destino.
          example: 8.5
        durationMin:
          type: integer
          description: Duración estimada a partir de la distancia.
          example: 15
        driver:
          $ref: "#/components/schemas/RideDriver"
        vehicle:
          $ref: "#/components/schemas/RideVehicle"
        conditions:
          type: array
          maxItems: 10
          items:
            type: string
          example: [No se admite desvíos, Máximo de espera 5 minutos]
    RideSummary:
      description: Viaje en el listado del dashboard (colección `viajes`), sin coordenadas.
      allOf:
        - $ref: "#/components/schemas/RideBase"
        - type: object
          required: [origin, destination]
          properties:
            origin:
              $ref: "#/components/schemas/PlaceSummary"
            destination:
              $ref: "#/components/schemas/PlaceSummary"
    PlaceSummary:
      type: object
      required: [label]
      properties:
        label:
          type: string
          example: Huaca del Dragón (Arco Iris)
        address:
          type: [string, "null"]
          example: La Esperanza, Trujillo
    RideDetail:
      description: Viaje con la ruta completa para el mapa (colección `viajes`).
      allOf:
        - $ref: "#/components/schemas/RideBase"
        - type: object
          required: [origin, destination, stops, createdAt]
          properties:
            origin:
              $ref: "#/components/schemas/Place"
            destination:
              $ref: "#/components/schemas/Place"
            stops:
              type: array
              maxItems: 5
              description: Paradas intermedias, en el orden del recorrido.
              items:
                $ref: "#/components/schemas/Place"
            createdAt:
              type: string
              format: date-time
    RideCreateRequest:
      type: object
      required:
        [direction, departureTime, origin, destination, pricePerSeat, seats]
      properties:
        direction:
          $ref: "#/components/schemas/RideDirection"
        departureTime:
          type: string
          format: date-time
          description: |
            Debe ser una fecha futura, con hora entre las 6:00 y las 23:00 de Perú
            (RN-12).
          example: "2026-10-09T12:30:00Z"
        origin:
          $ref: "#/components/schemas/Place"
        destination:
          $ref: "#/components/schemas/Place"
        stops:
          type: array
          maxItems: 5
          default: []
          description: Paradas intermedias, en el orden del recorrido.
          items:
            $ref: "#/components/schemas/Place"
        pricePerSeat:
          type: integer
          minimum: 1
          maximum: 10
          description: Aporte por plaza, en créditos (1 crédito = 1 sol); entre 1 y 10 (RN-13).
          example: 5
        seats:
          type: integer
          minimum: 1
          description: Plazas ofrecidas; como máximo, las plazas del vehículo (RN-11).
          example: 3
        conditions:
          type: array
          maxItems: 10
          default: []
          items:
            type: string
            minLength: 1
            maxLength: 120

    # --------------------------------------------------------------- Reservas
    ReserveRideRequest:
      type: object
      properties:
        seats:
          type: integer
          minimum: 1
          default: 1
          description: Plazas a reservar.